+++ assets/js/editor-blocks.js
@@ -2,6 +2,7 @@ import { createPressMathRenderer } from './math-render.js';
 import { createSafeHighlightFragment as createRuntimeSafeHighlightFragment } from './syntax-highlight.js';
 import { createEditorBlocksRuntime } from './editor-blocks-runtime.js';
 import { createEditorBlocksSessionRegistry } from './editor-blocks-session-registry.js';
+import { createEditorBlocksBlockActions } from './editor-blocks-block-actions.js';
 import { createEditorBlocksLayoutSession } from './editor-blocks-layout-session.js';
 import { createEditorBlocksBodySession } from './editor-blocks-body-session.js';
 import { createEditorBlocksStateController } from './editor-blocks-state.js';
@@ -31,7 +32,6 @@ import {
   createCaretSession,
   createInlineDomSession,
   editableText,
-  editableVisibleText,
   getEditableCaretTextOffset,
   getEditableSelectionOffsets,
   inlineMarkedDomRangeFromPointerEvent,
@@ -40,10 +40,8 @@ import {
   inlineRunsFromDom,
   insertCodeEditableTextAtSelection,
   insertPlainTextIntoEditable,
-  isEditableBackspaceAtEmptyStart,
   isEditableCaretOnEdgeLine,
   isEditableSelectionAtStart,
-  isEditableSelectionOnBlankLine,
   linkForTextRange,
   nodeContains,
   placeCaretAtEnd,
@@ -63,7 +61,6 @@ import {
 import {
   applyInlineLinkToRuns,
   applyInlineMathToRuns,
-  autofixMarkdownSourceBlock,
   convertListTailItemAfterEmptyToParagraph,
   defaultListItems,
   editableListItems,
@@ -75,17 +72,12 @@ import {
   inlineRangeText,
   inlineRun,
   insertInlineRunsAtRange,
-  isBlockEmptyForBackspace,
-  isMergeableListBlock,
   itemIndentLevel,
-  listBlockItems,
   listVisualMarkerLabels,
   makeBlankBlock,
   makeBlock,
   mergeFirstListItemIntoPreviousBlock,
   mergeListItemIntoPreviousItem,
-  mergeTextBlockIntoPrevious,
-  mergeTextBlockIntoPreviousList,
   normalizeEditableMarkdownText,
   normalizeTableAlignment,
   normalizeTableCellValue,
@@ -103,7 +95,6 @@ import {
   serializeMarkdownBlocks,
   splitBlankLineUnits,
   splitListItemsAtEmptyItem,
-  splitTextBlockIntoParagraph,
   summarizeListType,
   tableColumnCount,
   toggleInlineMarkOnRuns
@@ -261,20 +252,6 @@ export function createMarkdownBlocksEditor(root, options = {}) {
   const blockElements = () => Array.from(list.children).filter(el => el && el.classList && el.classList.contains('blocks-block'));
   const blockSessions = createEditorBlocksSessionRegistry();
 
-  const insertBlankBlock = (index = state.blocks.length, options = {}) => {
-    const { block, index: safeIndex } = blocksState.insertBlankBlock(index, options);
-    render();
-    if (options.command) {
-      queueMicrotask(() => {
-        blockSessions.focusFirstCommandItem(block.id);
-      });
-    } else if (options.focus !== false) {
-      focusBlockPrimaryEditable(block, 0);
-    }
-    emit();
-    return block;
-  };
-
   const focusBlockPrimaryEditable = (block, caretOffset = null) => {
     blockSessions.focusBlockPrimaryEditable(block, caretOffset);
   };
@@ -287,53 +264,35 @@ export function createMarkdownBlocksEditor(root, options = {}) {
     blockSessions.focusPreviousBlockEnd(index);
   };
 
-  const insertBlankBlockAfter = (index, editable = null, sync = null) => {
-    if (typeof sync === 'function') sync();
-    insertBlankBlock(Math.max(0, Math.min((Number(index) || 0) + 1, state.blocks.length)), { focus: true });
+  const setActive = (index, editable = null, sync = null) => {
+    blockSessions.setActive(index, editable, sync);
   };
 
-  const splitTextBlockAfterCaret = (event, block, index, editable = null) => {
-    if (!event || event.key !== 'Enter' || event.shiftKey || event.altKey || event.ctrlKey || event.metaKey || event.isComposing) return false;
-    if (!block || !['paragraph', 'quote', 'heading'].includes(block.type)) return false;
-    const offsets = getEditableSelectionOffsets(editable, caretSession);
-    if (!offsets || !offsets.collapsed) return false;
-    const currentText = editableVisibleText(editable);
-    if (offsets.start >= currentText.length || isEditableSelectionOnBlankLine(editable, caretSession)) return false;
-    const split = splitEditableTextAtSelection(editable, selectionSession);
-    if (!split.after) return false;
-    const nextBlocks = splitTextBlockIntoParagraph(block, split.before, split.after);
-    if (!nextBlocks) return false;
-    event.preventDefault();
-    blocksState.replaceBlocks(index, 1, nextBlocks);
-    render();
-    focusBlockPrimaryEditable(nextBlocks[1], 0);
-    emit();
-    return true;
-  };
+  const blockActions = createEditorBlocksBlockActions({
+    state,
+    blocksState,
+    blockSessions,
+    caretSession,
+    selectionSession,
+    getEditableSelectionOffsets,
+    focusBlockPrimaryEditable,
+    focusPreviousBlockEnd,
+    render: () => render(),
+    setActive,
+    emit,
+    queueTask: task => queueMicrotask(task)
+  });
 
-  const mergeTextBlockWithPreviousOnBackspace = (event, block, index, editable = null) => {
-    if (!event || event.key !== 'Backspace' || event.shiftKey || event.altKey || event.ctrlKey || event.metaKey || event.isComposing) return false;
-    if (!Number.isInteger(index) || index <= 0) return false;
-    if (!editable || !isEditableSelectionAtStart(editable, caretSession)) return false;
-    if (isBlockEmptyForBackspace(block)) return false;
-    const previous = state.blocks[index - 1] || null;
-    const previousItems = isMergeableListBlock(previous) ? listBlockItems(previous) : [];
-    const previousListItemIndex = previousItems.length - 1;
-    const merged = mergeTextBlockIntoPrevious(previous, block) || mergeTextBlockIntoPreviousList(previous, block);
-    if (!merged) return false;
-    event.preventDefault();
-    blocksState.replaceBlocks(index - 1, 2, [merged], {
-      pendingListFocus: merged.type === 'list' ? {
-        blockId: merged.id,
-        itemIndex: Number.isInteger(merged.focusItemIndex) ? merged.focusItemIndex : previousListItemIndex,
-        caretOffset: merged.focusCaretOffset
-      } : null
-    });
-    render();
-    if (merged.type !== 'list') focusBlockPrimaryEditable(merged, merged.focusCaretOffset);
-    emit();
-    return true;
-  };
+  const {
+    insertBlankBlock,
+    insertBlankBlockAfter,
+    splitTextBlockAfterCaret,
+    mergeTextBlockWithPreviousOnBackspace,
+    deleteBlockAt,
+    makeSplitListBlock,
+    removeEmptyBlockWithBackspace,
+    applySourceAutofix
+  } = blockActions;
 
   const clearNativeSelection = () => {
     selectionSession.clearSelection(root);
@@ -361,40 +320,8 @@ export function createMarkdownBlocksEditor(root, options = {}) {
     menuSession.closeInlineMenu(restoreFocus);
   };
 
-  const deleteBlockAt = (index) => {
-    const deleted = blocksState.deleteBlock(index);
-    if (!deleted) return;
-    render();
-    setActive(deleted.activeIndex);
-    emit();
-  };
-
-  const makeSplitListBlock = (block, items, after = '\n\n') => {
-    const data = block && block.data ? block.data : {};
-    return makeBlock('list', '', {
-      dirty: true,
-      listType: data.listType === 'ol' || data.listType === 'task' || data.listType === 'mixed' ? data.listType : 'ul',
-      items: Array.isArray(items) ? items.slice() : editableListItems(items).slice(),
-      after: after || '\n\n'
-    });
-  };
-
   const resetTransientBlockMenus = () => {
     blocksState.resetTransientMenus();
-  };
-
-  const removeEmptyBlockWithBackspace = (event, block, index, editable = null, sync = null) => {
-    if (!event || event.key !== 'Backspace' || event.shiftKey || event.altKey || event.ctrlKey || event.metaKey || event.isComposing) return false;
-    if (!Number.isInteger(index) || index <= 0) return false;
-    if (editable && !isEditableBackspaceAtEmptyStart(editable, selectionSession)) return false;
-    if (typeof sync === 'function') sync();
-    if (!isBlockEmptyForBackspace(block)) return false;
-    event.preventDefault();
-    blocksState.removeBlock(index);
-    render();
-    focusPreviousBlockEnd(index);
-    emit();
-    return true;
   };
 
   const actionMenuBoundaryLeft = () => {
@@ -418,16 +345,6 @@ export function createMarkdownBlocksEditor(root, options = {}) {
       const leftSpace = triggerRect ? triggerRect.right - boundaryLeft : menuRect.left - boundaryLeft;
       if (leftSpace < menuRect.width + 8) menu.classList.add('is-open-right');
     } catch (_) {}
-  };
-
-  const applySourceAutofix = (index) => {
-    const block = state.blocks[index];
-    const nextBlocks = autofixMarkdownSourceBlock(block);
-    if (!nextBlocks.length) return;
-    blocksState.replaceBlocks(index, 1, nextBlocks, { activeIndex: index });
-    render();
-    setActive(index);
-    emit();
   };
 
   const syncActiveEditable = () => {
@@ -454,10 +371,6 @@ export function createMarkdownBlocksEditor(root, options = {}) {
 
   const openMathEditorForBlock = (block, blockEl = null) => {
     blockSessions.openMathEditorForBlock(block, blockEl);
-  };
-
-  const setActive = (index, editable = null, sync = null) => {
-    blockSessions.setActive(index, editable, sync);
   };
 
   const activateEditableFromPointer = (index, editable, sync) => {
