+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.78",
-  "tag": "v3.4.78",
+  "version": "3.4.79",
+  "tag": "v3.4.79",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.77 <3.4.78"
+      ">=3.4.78 <3.4.79"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.77 first."
+    "message": "Update to v3.4.78 first."
   }
 }
